backend/pyproject.toml
[project]
name = "agentarium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "factory-boy>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "redis>=7.0.1",
    "types-django-filter>=25.2.0.20251010",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
warn_unused_configs = true
strict_optional = true

[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.base"

[dependency-groups]
dev = [
    "mypy>=1.11.0",
    "django-stubs>=5.0.2",
    "djangorestframework-stubs>=3.16.5",
]
